2010-09-08  Dean Jackson  <dino@apple.com>
        Unreviewed attempt to fix the Mac builds.

        * css/CSSParser.cpp:
        (WebCore::CSSParser::parseAnimationTimingFunction):

Canonical link: https://commits.webkit.org/57781@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@67036 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/css/CSSParser.cpp
@@ -2827,7 +2827,7 @@ PassRefPtr<CSSValue> CSSParser::parseAnimationTimingFunction()
         CSSParserValue* v = args->current();
         if (!validUnit(v, FInteger, m_strict))
             return 0;
-        numSteps = min(v->fValue, (double)INT_MAX);
+        numSteps = (int) min(v->fValue, (double)INT_MAX);
         if (numSteps < 1)
             return 0;
         v = args->next();
